requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp20210324.2
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
altair_data_server==0.4.1
altair_saver==0.5.0
altair_viewer==0.4.0
appnope==0.1.2
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow-cpp==5.0.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
async_generator==1.10
attrs==21.4.0
aws-c-cal==0.5.11
aws-c-common==0.6.2
aws-c-event-stream==0.2.7
aws-c-io==0.10.5
aws-checksums==0.1.11
aws-sdk-cpp==1.8.186
backcall==0.2.0
backports==1.0
backports.functools_lru_cache==1.6.4
backports.zoneinfo==0.2.1
base58==2.1.1
black==22.1.0
blas==1.0
bleach==4.1.0
blinker==1.4
boto3==1.21.4
botocore==1.24.5
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.18.1
ca-certificates==2022.4.26
cachetools==5.0.0
cairo==1.16.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cvxopt==1.2.7
cycler==0.11.0
cython==0.29.23
dataclasses==0.8
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
executing==0.8.2
flit-core==3.6.0
font-ttf-dejavu-sans-mono==2.37
font-ttf-inconsolata==3000
font-ttf-source-code-pro==2038
font-ttf-ubuntu==0.83
fontconfig==2.13.96
fonts-conda-ecosystem==1
fonts-conda-forge==1
fonttools==4.29.1
fpdf2==2.5.1
freetype==2.10.4
fribidi==1.0.10
frozenlist==1.3.0
gensim==4.0.1
gettext==0.19.8.1
gflags==2.2.2
giflib==5.2.1
gitdb==4.0.9
gitpython==3.1.27
glog==0.5.0
glove-py==0.2.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
graphite2==1.3.13
grpc-cpp==1.40.0
gspread==4.0.1
h11==0.13.0
harfbuzz==3.4.0
icu==69.1
idna==3.3
importlib-metadata==4.11.1
importlib_resources==5.4.0
intel-openmp==2020.2
intervaltree==3.1.0
ipykernel==6.9.1
ipython==8.0.1
ipython_genutils==0.2.0
ipywidgets==7.6.5
jbig==2.1
jedi==0.18.1
jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jpeg==9e
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter_client==7.1.2
jupyter_core==4.9.2
jupyterlab_pygments==0.1.2
jupyterlab_widgets==1.0.2
karateclub==1.2.3
kiwisolver==1.3.2
krb5==1.19.2
lcms2==2.12
lerc==2.2.1
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libcurl==7.79.1
libcxx==12.0.0
libdeflate==1.7
libedit==3.1.20191231
libev==4.33
libevent==2.1.10
libffi==3.4.2
libgfortran==3.0.1
libglib==2.70.2
libiconv==1.16
liblapack==3.9.0
libnghttp2==1.43.0
libpng==1.6.37
libprotobuf==3.18.0
libsodium==1.0.18
libssh2==1.10.0
libthrift==0.15.0
libtiff==4.3.0
libutf8proc==2.7.0
libuv==1.43.0
libwebp-base==1.2.2
libxml2==2.9.12
libzlib==1.2.11
llvm-openmp==13.0.1
lxml==4.8.0
lz4-c==1.9.3
markupsafe==2.1.0
matplotlib==3.5.1
matplotlib-base==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
mypy_extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
ncurses==6.3
nest-asyncio==1.5.4
networkx==2.6.3
nodejs==17.4.0
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
olefile==0.46
openjpeg==2.4.0
openssl==1.1.1o
orc==1.7.0
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandoc==2.17.1.1
pandocfilters==1.5.0
pango==1.50.5
parquet-cpp==1.5.1
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pcre==8.45
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.2
pip==21.2.4
pixman==0.40.0
platformdirs==2.5.1
pm4py==2.2.19.1
portpicker==1.5.0
prometheus_client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.18.0
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybind11==2.9.1
pycparser==2.21
pydeck==0.4.1
pydotplus==2.0.2
pygments==2.11.2
pygsp==0.5.1
pyjwt==2.3.0
pymfe==0.4.1
pympler==0.9
pyopenssl==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pysocks==1.7.1
python==3.8.12
python-dateutil==2.8.1
python-dotenv==0.20.0
python-graphviz==0.19.1
python-levenshtein==0.12.2
python-louvain==0.15
python-tzdata==2021.5
python_abi==3.8
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pyu2f==0.1.5
pyvis==0.1.9
pyyaml==6.0
pyzmq==22.3.0
re2==2021.09.01
readline==8.1.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
scikit-learn==1.0.2
scikit-multiflow==0.5.3
scipy==1.7.3
seaborn==0.11.0
selenium==4.1.0
send2trash==1.8.0
setuptools==58.0.4
six==1.15.0
smart_open==5.2.1
smmap==3.0.5
snappy==1.1.8
sniffio==1.2.0
sortedcontainers==2.4.0
sqlite==3.37.2
stack_data==0.2.0
statsmodels==0.13.2
streamlit==1.10.0
stringdist==1.0.9
sympy==1.9
tbb==2021.5.0
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.1.0
tk==8.6.11
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typed-ast==1.5.2
typing-extensions==4.1.1
typing_extensions==4.1.1
tzdata==2021e
tzlocal==4.1
unicodedata2==14.0.0
urllib3==1.26.8
validators==0.18.2
vega-cli==5.17.0
vega-lite-cli==4.17.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.5.2
wsproto==1.1.0
xz==5.2.5
yaml==0.2.5
yarl==1.7.2
zeromq==4.3.4
zipp==3.7.0
zlib==1.2.11
zstd==1.5.0
